Rotate HUD (right click menu) with canvas (map)
The HUD is a separate DOM element (a traditional <div>) that isn't connected to the <canvas> that renders the scene. Before, only the canvas was rotated, leaving all hud elements behind, usually off screen. Now, the hud element is given the same rotation as the canvas to place is correctly.

Note: Unlike the element that renders the scene, the hud is a classic html div, which means there isn't an easy property to set to rotate it. There are two ways to find the hud, either the way I've done it, or to find it by it's id `#hud`. I've opted to go through the `canvas` object that is used everywhere else, but I honestly don't know if there's a better way to do this.

Also note the `waitForHudToRender()` function. This exists because the canvas may render before the hud. Since this calls the `canvasReady` hook which eventually calls `applySettings()`, it's possible the hud doesn't yet exist and isn't ready to rotate. In order to rotate the hud, we must wait for it to render. This does not hold up the rest of the function, and after the initial render there is no delay